renamed section folder to pluralized
