+++ src/components/Sections/EdisonWhoSection.tsx
